pyproject.toml
[project]
name = "hoi-yo"
version = "0.1.0"
description = "Live AI Persona Agents for Hearts of Iron IV"
requires-python = ">=3.11"
dependencies = [
    "anthropic>=0.45.0",
    "click>=8.1",
    "fastapi>=0.110",
    "uvicorn[standard]>=0.27",
    "websockets>=12.0",
    "jinja2>=3.1",
    "watchdog>=4.0",
    "pydantic>=2.6",
    "httpx>=0.27",
    "pyjwt>=2.8",
    "bcrypt>=4.1",
    "aiosqlite>=0.19",
    "python-multipart>=0.0.5",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0",
    "pytest-asyncio>=0.23",
    "pytest-cov>=4.1",
    "ruff>=0.3",
]

[project.scripts]
hoi-yo = "src.cli:cli"

[build-system]
requires = ["setuptools>=68.0"]
build-backend = "setuptools.build_meta"

[tool.setuptools.packages.find]
include = ["src*"]

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]

[tool.ruff]
line-length = 100
